% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENCODE.r
\docType{data}
\name{ENCODE.human.mouse}
\alias{ENCODE.human.mouse}
\title{Normalized ENCODE microarray data for both human and mouse.}
\format{A data frame with 10309 rows and 26 columns}
\source{
Reproduced according to Gilad, Yoav, and Orna Mizrahi-Man. "A reanalysis of mouse ENCODE comparative gene expression data." F1000Research 4 (2015).
}
\usage{
data(ENCODE.human.mouse)
}
\description{
Normalized ENCODE microarray data for both human and mouse.
}
\keyword{datasets}
