% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{humanmouse}
\alias{humanmouse}
\title{Brain RNA-Seq data for both human and mouse.}
\format{A data frame with 15041 rows and 62 columns}
\source{
\url{http://web.stanford.edu/group/barres_lab/brainseq2/brainseq2.html}
}
\usage{
data(humanmouse)
}
\description{
Brain RNA-Seq data for both human and mouse.
}
\keyword{datasets}
