% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuantNorm.R
\name{QuantNorm}
\alias{QuantNorm}
\title{Conducting batch effect correction by quantile normalization}
\usage{
QuantNorm(dat, batch, method = "refB", iter = 1, logdat = TRUE,
  standardize = FALSE)
}
\arguments{
\item{dat}{The original p*n batch effect data with n subjects and p RNA-seq measurements.}

\item{batch}{The vector of length n indicating which batch the subjects belong to.}

\item{method}{Method for the quantile normalization. There are three options: "refB", "ref1" and "block".}

\item{iter}{Times of iteration for "refB" and "ref1" methods when no standardization is applied.}

\item{logdat}{Whether conducting log transformation to data or not.}

\item{standardize}{Whether conducting standardization [(dat - mean)/var] to data or not.}
}
\value{
The corrected 1-correlation matrix between subjects.
}
\description{
Conducting batch effect correction by quantile normalization
}
\examples{

library(rgl) #for 3D PCA display

data("humanmouse")

#Numbering the cells by cell types
celltype <- c(rep(1,8),rep(7,6),rep(1,12),rep(2,1),rep(3,5),rep(4,3),rep(5,2),rep(6,4),
              rep(1,2),rep(1,4),rep(2,2),rep(3,6),rep(4,2),rep(5,2),rep(6,3))

#Assigning the batch number that the 62 subjects belonging to.
batches <- c(rep(1,41),rep(2,21))

#Plot the 3D PCA for the uncorrected batch effect data
plot3d(princomp(1-cor(humanmouse,method='spearman'))$scores[,1:3], col=celltype)

#QuantNorm correction
ccc <- QuantNorm(humanmouse,batches,iter=10)
plot3d(princomp(ccc)$scores[,1:3], col=celltype, size=10)
}
